requirements.txt
astroid==2.4.2
click==7.1.2
colorama==0.4.3
comtypes==1.1.7
et-xmlfile==1.0.1
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.1
opencv-contrib-python==4.3.0.36
opencv-python==4.3.0.36
openpyxl==3.0.4
pandas==1.1.0
Pillow==7.2.0
pylint==2.5.3
pypiwin32==223
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyqt5-tools==5.15.0.1.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pyttsx3==2.90
pytz==2020.1
pywin32==228
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
